[no-relnote] Move fileinstaller to separate file
